smart_ai_farming_assistant/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
PyYAML>=6.0
requests>=2.28.0
pytest>=7.0.0
httpx>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0